2017/2/10 22:37 1.完善了控制器模块库的拖动连线功能 2.还没有实现添加数据库部分的功能，以及属性对话框部分的功能

+++ ROBOT/delaydialog.cpp
@@ -46,7 +46,7 @@ void DelayDialog::okButtonClicked()
 {
     QString str = ui->lineEdit->text();
     bool ok;
-    double time = str.toInt(&ok);
+    double time = str.toDouble(&ok);
     if(!ok){
         QMessageBox::warning(this, tr("delayDialog"), tr("请输入数字"));
         return;
